% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sw.R
\name{draw_sw}
\alias{draw_sw}
\title{Draw and export stepped-wedge cluster randomised trial design diagrams}
\usage{
draw_sw(design         = design_sw(),
        cp_contents    = design,
        row_names      = FALSE,
        col_names      = FALSE,
        colours,
        combine_rows   = FALSE,
        merge_cols     = FALSE,
        xlab           = "Time period",
        ylab           = "Cluster",
        cp_text_colour = "black",
        key_interventions,
        key_time_periods,
        make           = "print",
        filename       = "swcrt")
}
\arguments{
\item{design}{...}

\item{cp_contents}{...}

\item{row_names}{...}

\item{col_names}{...}

\item{colours}{...}

\item{combine_rows}{...}

\item{merge_cols}{...}

\item{xlab}{...}

\item{ylab}{...}

\item{cp_text_colour}{...}

\item{key_interventions}{...}

\item{key_time_periods}{...}

\item{make}{...}

\item{filename}{...}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{flextable}} in the slot \code{$table_design} containing
the constructed SW-CRT diagram.
\item A \code{\link{flextable}} in the slot \code{$table_interventions}
containing the constructed key summarising the various intervention states.
Will be \code{\link{NULL}} if input \code{key_interventions} is not
specified.
\item A \code{\link{flextable}} in the slot \code{$table_time_periods}
containing the constructed key summarising the various time periods.
Will be \code{\link{NULL}} if input \code{key_time_periods} is not
specified.
\item A \code{\link{list}} in the slot \code{$inputs} containing each of the
input variables.
}
}
\description{
...
}
\examples{
# The default is to draw a 'standard' SW-CRT design, with four clusters and
# five time periods
default_draw   <- draw_sw()
# Specify a more complex design, where all clusters begin in the intervention
# condition, there is an unequal number of clusters per wave, and there are
# two trailing extra time periods with all clusters in the intervention
# condition. Also modify the row and column names
complex_design <- design_sw(clusters_per_wave = c(1, 2, 3, 2, 0, 0),
                            row_names         = paste0("Cluster ", 1:8),
                            col_names         = paste0("Time period ", 1:6))
# Pass this to draw_sw(), producing .docx, .png, and .pptx files, merging
# columns, and combining rows
\dontrun{
complex_draw   <- draw_sw(complex_design, combine_rows = TRUE,
                          merge_cols = TRUE,
                          make = c("docx", "png", "pptx", "print"))
}
}
\seealso{
\code{\link{design_sw}}
}
\author{
Michael J Grayling (\email{michael.grayling@newcastle.ac.uk})
}
