% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface to package functionality}
\usage{
gui()
}
\description{
\code{gui()} runs an R Shiny web browser based graphical user interface for
running functions within \code{\link{swcrt}}.
}
\examples{
# Launch the graphical user interface
\dontrun{gui()}
}
\seealso{
\code{\link{opt_sw_norm}}, \code{\link{plot.swcrt_opt_sw_norm}}.
}
