% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swcrt.R
\docType{package}
\name{swcrt-package}
\alias{swcrt-package}
\alias{swcrt}
\title{Draw and export stepped-wedge cluster randomised trial design diagrams}
\description{
\strong{swcrt} is an R package that helps produce diagrams of stepped-wedge
cluster randomised trial (SW-CRT) designs, like those advocated for inclusion
when reporting SW-CRTs by
\href{https://doi.org/10.1136/bmj.k1614}{Hemming \emph{et al.} (2018)}. The
produced diagrams can be exported in .docx, .png, and .pptx formats.
}
\section{Getting started}{


You can install the latest development version of \strong{swcrt} from
\href{https://github.com/}{Github} with:

\code{devtools::install_github("mjg211/swcrt")}

An introductory example of how to make use of the package's core
functionality can be found \href{https://github.com/mjg211/swcrt}{here}.
More detailed support is available in the package vignette, which can be
accessed with \code{vignette("swcrt")}. For further help, please email
\email{michael.grayling@newcastle.ac.uk}.
}

\section{Details}{

\tabular{ll}{\strong{Package:} \tab swcrt\cr \strong{Type:} \tab Package\cr
\strong{Version:} \tab 0.1\cr \strong{Date:} \tab 2020-09-09\cr
\strong{License:} \tab MIT\cr}
The package exports two functions:
\itemize{
\item \code{\link{draw_sw}}: Draws SW-CRT designs that can be exported in a
variety of formats.
\item \code{\link{design_sw}}: A helper function to assist in producing
standard SW-CRT designs for subsequent use with \code{\link{draw_sw}}.
}
}

\section{References}{

Hemming K, Taljaard M, McKenzie JE, Hooper R, Copas A, Thompson JA,
Dixon-Woods M, Aldcroft A, Doussau A, Grayling M, Kristunas C, Goldstein CE,
Campbell MK, Girling A, Eldridge S, Campbell MJ, Lilford RJ, Weijer C,
Forbes AB, Grimshaw JM. Reporting of stepped wedge cluster randomised trials:
Extension of the CONSORT 2010 statement with explanation and elaboration.
\emph{BMJ} 2018;\strong{363}, k1614.
DOI: \href{https://doi.org/10.1136/bmj.k1614}{10.1136/bmj.k1614}.
PMID: \href{https://pubmed.ncbi.nlm.nih.gov/30413417/}{30413417}.
}

\author{
Michael J Grayling (\email{michael.grayling@newcastle.ac.uk})
}
