% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sw.R
\name{draw_sw}
\alias{draw_sw}
\title{Draw and export stepped-wedge cluster randomised trial design diagrams}
\usage{
draw_sw(design         = design_sw(),
        cp_contents    = design,
        row_names      = FALSE,
        col_names      = FALSE,
        colours,
        combine_rows   = FALSE,
        merge_cols     = FALSE,
        xlab           = "Time period",
        ylab           = "Cluster",
        cp_text_colour = "black",
        key_interventions,
        key_time_periods,
        make           = "print",
        filename       = "swcrt")
}
\arguments{
\item{design}{A SW-CRT design (i.e., information on the planned roll-out) to
draw a diagram for. Must be either a \code{\link{data.frame}} or a
\code{\link{matrix}}. Defaults to \code{\link{design_sw}()}.}

\item{cp_contents}{Values to add to the cluster-period cells in the produced
diagram; modify this when cluster-period cells should contain details
on something other than the intervention condition. Must be either a
\code{\link{data.frame}} or a \code{\link{matrix}}. Defaults to
\code{design}.}

\item{row_names}{A \code{\link{logical}} variable, indicating whether the row
names in the produced diagram should be taken from \code{design} or
constructed internally. Defaults to \code{\link{FALSE}}.}

\item{col_names}{A \code{\link{logical}} variable, indicating whether the
column names in the produced diagram should be taken from \code{design} or
constructed internally. Defaults to \code{\link{FALSE}}.}

\item{colours}{A named (see \code{\link{names}}) \code{\link{character}}
\code{\link{vector}}, giving the fill colours to add to the cluster-period
cells for each of the intervention conditions. Specified internally by
default.}

\item{combine_rows}{A \code{\link{logical}} variable, indicating whether to
combine rows that have identical sequences in the produced diagram. Defaults
to \code{\link{FALSE}}.}

\item{merge_cols}{A \code{\link{logical}} variable, indicating whether to
merge across columns when adjacent cluster-period cells have the same
intervention condition in the produced diagram. Defaults to
\code{\link{FALSE}}.}

\item{xlab}{A \code{\link{character}} string, giving the label for the x-axis
in the produced diagram. Defaults to \code{"Time period"}.}

\item{ylab}{A \code{\link{character}} string, giving the label for the y-axis
in the produced diagram. Defaults to \code{"Cluster"}.}

\item{cp_text_colour}{A \code{\link{character}} string, giving the colour to
use for the text in the cluster-period cells in the produced diagram.
Defaults to \code{"black"}.}

\item{key_interventions}{A named (see \code{\link{names}})
\code{\link{character}} \code{\link{vector}}, giving information to include
an intervention key. Not produced by default.}

\item{key_time_periods}{A named (see \code{\link{names}})
\code{\link{character}} \code{\link{vector}}, giving information to include
a time period key. Not produced by default.}

\item{make}{A \code{\link{character}} \code{\link{vector}}, giving the
(potentially multiple) types of output (export) that are desired. Can include
\code{"print"} (view within R), "docx" (save a .docx MS Word file), "png"
(save a .png file), and "docx" (save a .pptx MS PowerPoint file). Defaults to
\code{"print"}.}

\item{filename}{A \code{\link{character}} string, giving the filename to use
for any files that are to be saved. Defaults to \code{"swcrt"}.}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{flextable}} in the slot \code{$table_design} containing
the constructed SW-CRT diagram.
\item A \code{\link{flextable}} in the slot \code{$table_interventions}
containing the constructed key summarising the various intervention states.
Will be \code{\link{NULL}} if input \code{key_interventions} is not
specified.
\item A \code{\link{flextable}} in the slot \code{$table_time_periods}
containing the constructed key summarising the various time periods.
Will be \code{\link{NULL}} if input \code{key_time_periods} is not
specified.
\item A \code{\link{list}} in the slot \code{$inputs} containing each of the
input variables.
}
}
\description{
\code{\link{draw_sw}} draws stepped-wedge cluster randomised trial design
diagrams that can be exported in .docx, .png, and .pptx formats.
}
\details{
More details on each of the input variables can be found in the package
\code{\link{vignette}}.
}
\examples{
# The default is to draw a 'standard' SW-CRT design, with four clusters and
# five time periods
default_draw   <- draw_sw()
# Specify a more complex design, where all clusters begin in the intervention
# condition, there is an unequal number of clusters who switch per time
# period, and there are two trailing extra time periods with all clusters in
# the intervention condition. Also modify the row and column names
complex_design <- design_sw(clusters_per_time_period = c(1, 2, 3, 2, 0, 0),
                            row_names                = paste("Cluster", 1:8),
                            col_names                = paste("Time period",
                                                             1:6))
# Pass this to draw_sw(), producing .docx, .png, and .pptx files, merging
# columns, and combining rows
\dontrun{
complex_draw   <- draw_sw(complex_design, combine_rows = TRUE,
                          merge_cols = TRUE,
                          make = c("docx", "png", "pptx", "print"))
}
}
\seealso{
\code{\link{design_sw}}
}
\author{
Michael J Grayling (\email{michael.grayling@newcastle.ac.uk})
}
